refactor(api): split and host api/index.js functions in *.api.js
