+++ src/api/index.js
@@ -158,138 +158,7 @@ export const v3 = {
   },
 };
 
-export const fetchAuthUser = accessToken => v3.getJson('/user', accessToken);
-
-export const fetchAuthUserOrgs = accessToken =>
-  v3.getJson('/user/orgs', accessToken);
-
-export const fetchUser = (user, accessToken) =>
-  v3.getJson(`/users/${user}`, accessToken);
-
-export const fetchUserOrgs = (user, accessToken) =>
-  v3.getJson(`/users/${user}/orgs`, accessToken);
-
-export const fetchUserEvents = (user, accessToken) =>
-  v3.getJson(`/users/${user}/received_events?per_page=100`, accessToken);
-
-export const fetchReadMe = (user, repository, accessToken) =>
-  v3.getHtml(`/repos/${user}/${repository}/readme?ref=master`, accessToken);
-
-export const fetchOrg = (orgName, accessToken) =>
-  v3.getJson(`/orgs/${orgName}`, accessToken);
-
-export const fetchOrgMembers = (orgName, accessToken) =>
-  v3.getJson(`/orgs/${orgName}/members`, accessToken);
-
-export const fetchPostIssueComment = (
-  body,
-  owner,
-  repoName,
-  issueNum,
-  accessToken
-) =>
-  v3.postJson(
-    `/repos/${owner}/${repoName}/issues/${issueNum}/comments`,
-    accessToken,
-    { body }
-  );
-
-export const fetchEditIssue = (
-  owner,
-  repoName,
-  issueNum,
-  editParams,
-  updateParams,
-  accessToken
-) =>
-  v3.patch(
-    `/repos/${owner}/${repoName}/issues/${issueNum}`,
-    accessToken,
-    editParams
-  );
-
-export const fetchChangeIssueLockStatus = (
-  owner,
-  repoName,
-  issueNum,
-  currentStatus,
-  accessToken
-) =>
-  v3[currentStatus ? 'delete' : 'put'](
-    `/repos/${owner}/${repoName}/issues/${issueNum}/lock`,
-    accessToken
-  );
-
-export const fetchSearch = (type, query, accessToken, params = '') =>
-  v3.getJson(`/search/${type}?q=${query}${params}`, accessToken);
-
-export const fetchNotifications = (participating, all, accessToken) =>
-  v3.getJson(
-    `/notifications?participating=${participating}&all=${all}`,
-    accessToken
-  );
-
-export const fetchMarkNotificationAsRead = (notificationID, accessToken) =>
-  v3.patch(`/notifications/threads/${notificationID}`, accessToken);
-
-export const fetchMarkRepoNotificationAsRead = (repoFullName, accessToken) =>
-  v3.put(`/repos/${repoFullName}/notifications`, accessToken);
-
-export const fetchChangeStarStatusRepo = (owner, repo, starred, accessToken) =>
-  v3[starred ? 'delete' : 'put'](`/user/starred/${owner}/${repo}`, accessToken);
-
-export const fetchForkRepo = (owner, repo, accessToken) =>
-  v3.post(`/repos/${owner}/${repo}/forks`, accessToken);
-
-export const fetchStarCount = (owner, accessToken) =>
-  v3.count(`/users/${owner}/starred`, accessToken);
-
-export const isWatchingRepo = (url, accessToken) => v3.head(url, accessToken);
-
-export const watchRepo = (owner, repo, accessToken) =>
-  v3.put(`/repos/${owner}/${repo}/subscription`, accessToken, {
-    subscribed: true,
-  });
-
-export const unWatchRepo = (owner, repo, accessToken) =>
-  v3.delete(`/repos/${owner}/${repo}/subscription`, accessToken);
-
-export const fetchChangeFollowStatus = (user, isFollowing, accessToken) =>
-  v3[isFollowing ? 'delete' : 'put'](`/user/following/${user}`, accessToken);
-
-export const fetchDiff = (url, accessToken) => v3.getDiff(url, accessToken);
-
-export const fetchMergeStatus = (repo, issueNum, accessToken) =>
-  v3.get(`/repos/${repo}/pulls/${issueNum}/merge`, accessToken);
-
-export const fetchMergePullRequest = (
-  repo,
-  issueNum,
-  commitTitle,
-  commitMessage,
-  mergeMethod,
-  accessToken
-) =>
-  v3.put(`/repos/${repo}/pulls/${issueNum}/merge`, accessToken, {
-    commit_title: commitTitle,
-    commit_message: commitMessage,
-    merge_method: mergeMethod,
-  });
-
-export const fetchSubmitNewIssue = (
-  owner,
-  repo,
-  issueTitle,
-  issueComment,
-  accessToken
-) =>
-  v3.postJson(`/repos/${owner}/${repo}/issues`, accessToken, {
-    title: issueTitle,
-    body: issueComment,
-  });
-
-// Auth
-const authParameters = (code, state) => ({
+export const authParameters = (code, state) => ({
   method: METHOD.POST,
   headers: {
     Accept: 'application/json',
@@ -302,13 +171,3 @@ const authParameters = (code, state) => ({
     state,
   }),
 });
-
-export async function fetchAccessToken(code, state) {
-  const GITHUB_OAUTH_ENDPOINT = 'https://github.com/login/oauth/access_token';
-  const response = await fetch(
-    GITHUB_OAUTH_ENDPOINT,
-    authParameters(code, state)
-  );
-
-  return response.json();
-}
