+++ src/search/search.api.js
@@ -1,0 +1,4 @@
+import { v3 } from 'api';
+
+export const fetchSearch = (type, query, accessToken, params = '') =>
+  v3.getJson(`/search/${type}?q=${query}${params}`, accessToken);
